requirements.txt
appdirs==1.4.4
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
Django==3.2
django-filter==2.4.0
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
factory-boy==3.2.0
Faker==8.1.2
filelock==3.0.12
html5lib==1.1
idna==2.10
itypes==1.2.0
Jinja2==3.0.0
lxml==4.6.3
MarkupSafe==2.0.0
oauthlib==3.1.0
Pillow==8.2.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.2.1
sqlparse==0.4.1
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.6
virtualenvwrapper-win==1.2.6
webencodings==0.5.1
